Rename test files back
